Always show population color legend in side bar
